review: Clean up statefun-smoke-e2e-embedded
1. Move all embedded fn code to a specific package
2. Remove redundant fat from POM
3. No need to pre-generate user-level commands (can just reuse the ones
in driver, since the embedded module depends on driver and we will be
building a single jar anyways).
4. Rename some classes / add comments for sake of clarity
