+++ statefun-e2e-tests/statefun-smoke-e2e-driver/src/main/java/org/apache/flink/statefun/e2e/smoke/Types.java
@@ -18,24 +18,26 @@
 
 package org.apache.flink.statefun.e2e.smoke;
 
+import org.apache.flink.statefun.e2e.smoke.generated.Commands;
 import org.apache.flink.statefun.e2e.smoke.generated.SourceCommand;
 import org.apache.flink.statefun.e2e.smoke.generated.VerificationResult;
 import org.apache.flink.statefun.sdk.TypeName;
 import org.apache.flink.statefun.sdk.reqreply.generated.TypedValue;
 
-final class Types {
+public final class Types {
   private Types() {}
 
   public static final TypeName SOURCE_COMMANDS_TYPE =
       TypeName.parseFrom(Constants.NAMESPACE + "/sourceCommand");
   public static final TypeName VERIFICATION_RESULT_TYPE =
       TypeName.parseFrom(Constants.NAMESPACE + "/verificationResult");
+  public static final TypeName COMMANDS_TYPE = TypeName.parseFrom("statefun.smoke.e2e/commands");
 
-  static boolean isTypeOf(TypedValue value, TypeName type) {
+  public static boolean isTypeOf(TypedValue value, TypeName type) {
     return value.getTypename().equals(type.canonicalTypenameString());
   }
 
-  static TypedValue packSourceCommand(SourceCommand sourceCommand) {
+  public static TypedValue packSourceCommand(SourceCommand sourceCommand) {
     return TypedValue.newBuilder()
         .setTypename(SOURCE_COMMANDS_TYPE.canonicalTypenameString())
         .setHasValue(true)
@@ -43,7 +45,7 @@ final class Types {
         .build();
   }
 
-  static SourceCommand unpackSourceCommand(TypedValue typedValue) {
+  public static SourceCommand unpackSourceCommand(TypedValue typedValue) {
     if (!isTypeOf(typedValue, SOURCE_COMMANDS_TYPE)) {
       throw new IllegalStateException("Unexpected TypedValue: " + typedValue);
     }
@@ -54,7 +56,34 @@ final class Types {
     }
   }
 
-  static VerificationResult unpackVerificationResult(TypedValue typedValue) {
+  public static TypedValue packCommands(Commands commands) {
+    return TypedValue.newBuilder()
+        .setTypename(COMMANDS_TYPE.canonicalTypenameString())
+        .setHasValue(true)
+        .setValue(commands.toByteString())
+        .build();
+  }
+
+  public static Commands unpackCommands(TypedValue typedValue) {
+    if (!isTypeOf(typedValue, COMMANDS_TYPE)) {
+      throw new IllegalStateException("Unexpected TypedValue: " + typedValue);
+    }
+    try {
+      return Commands.parseFrom(typedValue.getValue());
+    } catch (Exception e) {
+      throw new RuntimeException("Unable to parse Commands from TypedValue.", e);
+    }
+  }
+
+  public static TypedValue packVerificationResult(VerificationResult verificationResult) {
+    return TypedValue.newBuilder()
+        .setTypename(VERIFICATION_RESULT_TYPE.canonicalTypenameString())
+        .setHasValue(true)
+        .setValue(verificationResult.toByteString())
+        .build();
+  }
+
+  public static VerificationResult unpackVerificationResult(TypedValue typedValue) {
     if (!isTypeOf(typedValue, VERIFICATION_RESULT_TYPE)) {
       throw new IllegalStateException("Unexpected TypedValue: " + typedValue);
     }
